refactor: tracks route code changes
